+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [refactor: optimize log](https://github.com/Tencent/spring-cloud-tencent/pull/1700)
 - [feat:support setting load balancing strategy per service.](https://github.com/Tencent/spring-cloud-tencent/pull/1701)
 - [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1702)
+- [feat:support traffic mirroring.](https://github.com/Tencent/spring-cloud-tencent/pull/1703)